requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.2.0
chardet==4.0.0
Cython==0.29.32
discord.py==1.7.3
idna==3.4
multidict==6.0.4
numpy==1.21.6
pandas==1.3.5
Pillow==9.4.0
pkg-resources==0.0.0
pyboy==1.5.3
PySDL2==0.9.14
pysdl2-dll==2.26.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
typing-extensions==4.4.0
yarl==1.8.2